+++ src/gearbox/routers/match_form.py
@@ -44,6 +44,15 @@ async def get_match_form(
     presigned_url = bucket_utils.get_presigned_url(request, config.S3_BUCKET_MATCH_FORM_KEY_NAME, params, "get_object")
     return JSONResponse(presigned_url, status.HTTP_200_OK)
 
+@mod.get("/important-questions", dependencies=[ Depends(auth.authenticate)], status_code=status.HTTP_200_OK)
+async def get_match_form(
+    request: Request,
+    session: Session = Depends(deps.get_session)
+):
+    params = []
+    presigned_url = bucket_utils.get_presigned_url(request, config.S3_BUCKET_IMPORTANT_QUESTIONS_KEY_NAME, params, "get_object")
+    return JSONResponse(presigned_url, status.HTTP_200_OK)
+
 @mod.post("/update-match-form", dependencies=[ Depends(auth.authenticate), Depends(admin_required)])
 async def update_sae(
     body: MatchForm,
